Extract CardTag into new file
